@@ -20,8 +20,8 @@ import org.apache.shardingsphere.sharding.route.strategy.ShardingStrategy;
 import org.apache.shardingsphere.sharding.rule.ShardingRule;
 import org.springframework.stereotype.Component;
 
-import javax.annotation.PostConstruct;
-import javax.annotation.Resource;
+import jakarta.annotation.PostConstruct;
+import jakarta.annotation.Resource;
 import javax.sql.DataSource;
 import java.sql.SQLException;
 import java.util.Collections;
@@ -34,7 +34,7 @@ import java.util.stream.Collectors;
  * 动态分表管理
  * 首次加载项目时需要更新分表配置
  */
-@Component
+// @Component
 @Slf4j
 public class DynamicShardingManager {
 
