@@ -6,7 +6,7 @@ import com.zzx.zzxpicturebackend.service.PictureService;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 
 import java.util.List;
 
