@@ -4,7 +4,7 @@ import com.zzx.zzxpicturebackend.service.SpaceAnalyzeService;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 
 import static org.junit.jupiter.api.Assertions.*;
 
